move files and fix initial state stuff
